# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "РЕЙС 400 — API",
    "version": "2.1.0",
    "description": "Server-authoritative v2 ShiftRun and separate practice XP / weekly SP alongside legacy v1. Synthetic training, not for employment decisions. See docs/API-V2.md."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/session": {
      "post": {
        "summary": "Create synthetic profile or reuse current session",
        "security": [],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "201": {
            "description": "Created; response includes profile bootstrap or authoritative run."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "crew": {
                    "type": "string",
                    "enum": [
                      "msk-1",
                      "msk-2",
                      "spb-1",
                      "spb-2"
                    ]
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
    "/api/bootstrap": {
      "get": {
        "summary": "Profile, catalogue, achievements, challenge and notices",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        }
      }
    },
    "/api/profile": {
      "patch": {
        "summary": "Change synthetic crew",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "crew": {
                    "type": "string",
                    "enum": [
                      "msk-1",
                      "msk-2",
                      "spb-1",
                      "spb-2"
                    ]
                  }
                },
                "required": [
                  "crew"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete own profile and all its activity",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "confirm": {
                    "const": true
                  }
                },
                "required": [
                  "confirm"
                ]
              }
            }
          }
        }
      }
    },
    "/api/profile/export": {
      "get": {
        "summary": "Export own data without credentials",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "Revoke current session",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {},
                "required": []
              }
            }
          }
        }
      }
    },
    "/api/leaderboard": {
      "get": {
        "summary": "Top 50 within crew, depot or company",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "crew",
                "depot",
                "company"
              ],
              "default": "crew"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        }
      }
    },
    "/api/notices/read": {
      "post": {
        "summary": "Read own notices",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "ids": {
                    "type": "array",
                    "maxItems": 100,
                    "items": {
                      "type": "string",
                      "maxLength": 249
                    }
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "summary": "Start one scored or practice attempt",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "201": {
            "description": "Created; response includes profile bootstrap or authoritative run."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "scenarioId": {
                    "type": "string"
                  },
                  "practice": {
                    "type": "boolean"
                  },
                  "requestId": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 64,
                    "pattern": "^[a-zA-Z0-9-]+$"
                  }
                },
                "required": [
                  "scenarioId",
                  "requestId"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/v1/results": {
      "get": {
        "summary": "Read-only HR/LMS result events",
        "security": [
          {
            "IntegrationKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        }
      }
    },
    "/api/integrations/v1/profiles": {
      "get": {
        "summary": "Read-only HR/points-accounting profile snapshots",
        "security": [
          {
            "IntegrationKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "maxLength": 36
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. See docs/API.md for complete response fields, scoring and pagination semantics."
          },
          "default": {
            "description": "JSON error: {error:{code,message}}. See docs/API.md for error codes."
          }
        }
      }
    },
    "/api/v2/catalog": {
      "get": {
        "summary": "Published summaries without hidden graphs",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. Authoritative state; see docs/API-V2.md.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {},
                        "required": [],
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "JSON error. DEADLINE_EXPIRED 409 can include a committed timeout. Retry the same requestId after uncertain delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/runs": {
      "post": {
        "summary": "Create pinned briefing; reserve competition attempt only at begin",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success. Authoritative state; see docs/API-V2.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftPublic"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. DEADLINE_EXPIRED 409 can include a committed timeout. Retry the same requestId after uncertain delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftStart"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/motivation": {
      "get": {
        "summary": "Separated evidence, XP, SP, goal, review and achievements",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success. Authoritative state; see docs/API-V2.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Motivation"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. DEADLINE_EXPIRED 409 can include a committed timeout. Retry the same requestId after uncertain delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/motivation/preferences": {
      "post": {
        "summary": "Set goal and reminders, without pausing a run",
        "security": [
          {
            "Session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Reis-Client",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Authoritative state; see docs/API-V2.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Motivation"
                }
              }
            }
          },
          "default": {
            "description": "JSON error. DEADLINE_EXPIRED 409 can include a committed timeout. Retry the same requestId after uncertain delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotivationPreferences"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ShiftStart": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "pattern": "^[a-zA-Z0-9-]+$"
          },
          "scenarioId": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "training",
              "assessment"
            ]
          },
          "timingPolicyId": {
            "type": "string",
            "enum": [
              "standard",
              "extended",
              "untimed"
            ]
          },
          "variantId": {
            "type": "string"
          },
          "serviceClass": {
            "type": "string",
            "enum": [
              "standard",
              "comfort",
              "business",
              "first"
            ]
          },
          "competitionSlotId": {
            "const": "shift"
          }
        },
        "required": [
          "requestId",
          "scenarioId",
          "mode",
          "timingPolicyId"
        ],
        "additionalProperties": false
      },
...
      "ShiftError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          },
          "run": {
            "$ref": "#/components/schemas/ShiftPublic"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "reis_session"
      },
      "IntegrationKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}